pyproject.toml
[tool.poetry]
name = "trafficpredictions"
version = "0.1.0"
description = "Project on building prediction models for traffic in Slovenia"
authors = ["Patrik Praprotnik <praprotnik.patrik@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.2"
numpy = "^1.26.4"
python-dotenv = "^1.0.1"
pyproj = "^3.6.1"
pandas = "^2.2.2"
holidays = "^0.49"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
